rename pretraining dataset
